refactor: 1. move TodoWrapper to App; 2. export type Todo for reuse
