+++ src/components/Todo.tsx
@@ -1,0 +1,5 @@
+export type Todo = {
+  id: string;
+  task: string;
+  isEditing: boolean;
+};
